refactor(perf): split /youtube + /prototypes into server shell + motion island
33rd + 34th in the pattern series.
